Number Verifier 4.0.1.1 / 2022 - separate reporting out of NumberVerifierMain - use settings to determine what to report to the user
