+++ Number Verifier/Reporter/IMessageFilter.cs
@@ -1,0 +1,7 @@
+﻿namespace Sdl.Community.NumberVerifier.Reporter
+{
+	public interface IMessageFilter
+	{
+		bool IsAllowed(string message);
+	}
+}
